pyproject.toml
[tool.poetry]
name = "new-admin-panel-sprint-3"
version = "0.1.0"
description = ""
authors = ["Valerii-Denisov <klerk919191@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
flake8 = "6.1.0"
python-dotenv = "^1.0.1"
psycopg = {version = "3.1.18", extras = ["binary"]}
elasticsearch = "^8.15.0"
pydantic = "^2.8.2"
pyyaml = "^6.0.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
